app-2/pyproject.toml
[tool.poetry]
name = "app-1"
version = "0.1.0"
description = ""
authors = ["Brandon Braner <brandon.braner@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
opentelemetry-instrumentation = "^0.35b0"
opentelemetry-distro = "^0.35b0"
opentelemetry-exporter-otlp-proto-grpc = "^1.14.0"
requests = "^2.28.1"
opentelemetry-instrumentation-aws-lambda = "^0.35b0"
opentelemetry-instrumentation-dbapi = "^0.35b2"
opentelemetry-instrumentation-logging = "^0.35b0"
opentelemetry-instrumentation-sqlite3 = "^0.35b2"
opentelemetry-instrumentation-urllib = "^0.35b0"
opentelemetry-instrumentation-wsgi = "^0.35b0"
opentelemetry-instrumentation-fastapi = "^0.35b0"
opentelemetry-instrumentation-grpc = "^0.35b0"
opentelemetry-instrumentation-requests = "^0.35b0"
opentelemetry-instrumentation-urllib3 = "^0.35b0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
